gpu linear system

+++ src/main/native/src/settings.h
@@ -14,8 +14,8 @@ struct Settings
     bool DEBUG_SOLVER_CONVERGENCE   = false;        // 8
     bool DEBUG_CHECK_ARG            = false;        // 9
 
-    size_t GRID_SIZE        = 1;    // 10
-    size_t BLOCK_SIZE       = 512;  // 11
+    unsigned int GRID_SIZE        = 1;    // 10
+    unsigned int BLOCK_SIZE = 512; // 11
 
     double CU_SOLVER_LWORK_FACTOR   = 1.0;      // 21
     double CU_SOLVER_EPSILON        = 10.0;     // 22
